Criado hook para autenticação
