+++ src/Ui/src/hooks/index.ts
@@ -4,3 +4,4 @@ export * from "./useLocation"
 export * from "./useAuth"
 export * from "./usePersistentState"
 export * from "./useFilterUsers"
+export * from "./useAuthRequest"
